+++ src/androidTest/kotlin/us/frollo/frollosdk/logging/NetworkLoggerTest.kt
@@ -31,8 +31,8 @@ import us.frollo.frollosdk.testutils.trimmedPath
 
 class NetworkLoggerTest : BaseAndroidTest() {
 
-    override fun initSetup() {
-        super.initSetup()
+    override fun initSetup(daOAuth2Login: Boolean) {
+        super.initSetup(daOAuth2Login)
 
         preferences.loggedIn = true
         preferences.encryptedAccessToken = keystore.encrypt("ExistingAccessToken")