+++ src/main/kotlin/us/frollo/frollosdk/network/NetworkHelper.kt
@@ -36,7 +36,7 @@ internal class NetworkHelper(private val appInfo: AppInfo) {
         internal const val HEADER_BACKGROUND = "X-Background"
         internal const val HEADER_OTP = "X-User-Otp"
         internal const val HEADER_SESSION_ID = "X-Session-Id"
-        internal const val API_VERSION = "2.15"
+        internal const val API_VERSION = "2.16"
     }
 
     internal val bundleId: String